// src/app/model/user.model.ts
export class User {
    id: number;
    fname: string;
    lname: string;
    email: string;
    number: number;
    age: number;
    country: string;
    states: string;
    address: string;
    haddress1: string;
    haddress2: string;
    caddress1: string;
    caddress2: string;
    subscribe: string;
}